**Summary:**  
Improvement to the execution model of previous Loader, with the usage of multiple green threads for increasing efficiency. 

**Issue Reference(s):**  
Fixes #419 

To boost the computation, removed the vector allocation from `&[DataLoaderRequest]` which performed following sorting as mentioned below.
```rust
let mut keys = keys.to_vec(); 
// expensive as we clone and perform comparison on that.
keys.sort_by(|a, b| a.to_request().url().cmp(b.to_request().url())); 
```
and some single threaded iteration were modified to support operating with multiple green thread which internally spawned on the default runtime of tokio
``` rust 
let urls = keys[1..].iter().map(|key: &DataLoaderRequest| {
        let key_request_result = key.to_request();
        tokio::task::spawn(async move {
          ...
          ...
          //move some IO computation over here 
        })
      });
```
this should theoretically improve the loader performance as we are offloading work to multiple green threads

**Build & Testing:**

- [x] I ran `cargo build` successfully.
- [ ] I have run `./lint.sh --mode=fix` to fix all linting issues raised by `./lint.sh --mode=check`.

**Checklist:**

- [x] I have performed a self-review of my own code.