@artech-git 2023-11-07
> @artech-git thanks! Are you able to observe any improvement in performance?

some benchmarks result here (sorting procedure for keys eliminated and no explicit heap allocations)
from these I bought it to about ~ 18.768% only! 
without batching... 
```bash
 wrk -t4 -c100 -d5s -s ./ci-benchmark/wrk.lua http://localhost:8000/graphql
Running 5s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.89ms    3.53ms  26.62ms   66.54%
    Req/Sec     3.68k   464.98     7.20k    95.07%
  74325 requests in 5.10s, 381.49MB read
Requests/sec:  14583.61
Transfer/sec:     74.85MB
```
with batching
```bash
wrk -t4 -c100 -d5s -s ./ci-benchmark/wrk.lua http://localhost:8000/graphql
Running 5s test @ http://localhost:8000/graphql
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.77ms    1.74ms  16.72ms   70.46%
    Req/Sec     4.35k   171.60     4.69k    81.86%
  88327 requests in 5.10s, 453.35MB read
Requests/sec:  17320.32
Transfer/sec:     88.90MB
```
Iam little surprised to the improvements here, so I request you to verify the same, meanwhile I will run more test cases and revisions.