@meskill 2023-11-07
Looking in the changes I have doubts that the changes actually improved the performance despite the numbers in benchmarks.

There could be done deeper investigations with tracing etc to figure out the cause of the performance downsides since when running tc server we are proxying the requests to another server that could affect the result numbers, right?