@artech-git 2023-11-07
> > > Looking in the changes I have doubts that the changes actually improved the performance despite the numbers in benchmarks.
> > > There could be done deeper investigations with tracing etc to figure out the cause of the performance downsides since when running tc server we are proxying the requests to another server that could affect the result numbers, right?
> > 
> > 
> > yup exactly since tailcall is just a relay to a actual graphql server ! yeah I agree, but since I eliminated the **sorting after allocation of slice of DataLoaderRequest**'s, It is natural to pick up more speed since we are only iterating over slice here and not by dereferencing a pointer to heap of DataLoaderRequest(as I understood it is not nessecary as of now), but yeah it is very important though to have a precise benchmarks for multiple sections in the modified commit as it will enables to look at a better picture and plan for changes accordingly😊.
> 
> btw do you actually get many entries for keys in single call in your tests? I think the `delay: 0` in the issue example may effect this

yeah I was thinking about the same here, need some help with that. 🤔