@artech-git 2023-11-08
@meskill the facing trouble with logging feature of `log` here🤔, tried to enable that using feature flag of of `env_logger` didn't worked, can you help me with that ?