@meskill 2023-11-08
> @meskill the facing trouble with logging feature of `log` here🤔, tried to enable that using feature flag of of `env_logger` didn't worked, can you help me with that ?

could you provide more details on what have you tried?
if you're trying to enable additional logs you could use flag `--log-level` when running tc
another option is to use just builtin macroses like println and dbg if you want to test something quick