@artech-git 2023-11-08
> the simplest way to check this is to add log with keys.len() and run the tests to see what is logged 😀

`keys.len()` is 1 for `{ delay : 0 }`